Added parseFloat to script

+++ js/script.js
@@ -6,8 +6,8 @@
 /* FUnction */
 function calculatePay() {
   /* Input */
-  const hoursWorked = parseInt(document.getElementById("hoursWorked").value)
-  const wage = parseInt(document.getElementById("hourlyRate").value)
+  const hoursWorked = parseFloat(document.getElementById("hoursWorked").value)
+  const wage = parseFloat(document.getElementById("hourlyRate").value)
   const TAX_RATE = 0.18
   const pay = hoursWorked * wage * (1 - TAX_RATE)
   const taxes = hoursWorked * wage * TAX_RATE
